+++ LibraryApp_WebAPI/Models/Author.cs
@@ -9,7 +9,7 @@ namespace LibraryApp_WebAPI.Models
     public class Author
     {
         [Key]
-        public int AuthorId { get; set; }
+        public long AuthorId { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public DateTime DateOfBirth { get; set; }
