+++ LibraryApp_WebAPI/Models/BookAuthor.cs
@@ -11,13 +11,13 @@ namespace LibraryApp_WebAPI.Models
     {
         
         [ForeignKey("Author")]
-        public int AuthorId { get; set; }
+        public long AuthorId { get; set; }
         public Author Author { get; set; }
 
 
         
         [ForeignKey("Book")]
-        public int BookId { get; set; }
+        public long BookId { get; set; }
         public Book Book { get; set; }
